// frontend/src/api/generated/zod.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

import * as z from 'zod';

/**
 * HealthResponse
 */
export const zHealthResponse = z.object({
    status: z.string()
});

/**
 * TodoRead
 */
export const zTodoMutationResponseTodoRead = z.object({
    id: z.uuid(),
    title: z.string(),
    completed: z.boolean(),
    created_at: z.iso.datetime(),
    updated_at: z.iso.datetime(),
    version: z.int()
});

/**
 * TodoMutationResponse
 */
export const zTodoMutationResponse = z.object({
    todo: zTodoMutationResponseTodoRead,
    txid: z.int()
});

/**
 * ProblemDetail
 */
export const zValidationError = z.object({
    type: z.string(),
    title: z.string(),
    status: z.int(),
    detail: z.string(),
    instance: z.string(),
    errors: z.record(z.string(), z.array(z.string())).nullish().default(null)
});

/**
 * TodoCreate
 */
export const zTodoCreate = z.object({
    id: z.uuid(),
    title: z.string().min(1).max(200)
});

/**
 * ProblemDetail
 */
export const zHttpError = z.object({
    type: z.string(),
    title: z.string(),
    status: z.int(),
    detail: z.string(),
    instance: z.string(),
    errors: z.record(z.string(), z.array(z.string())).nullish().default(null)
});

/**
 * TodoPatch
 */
export const zTodoPatch = z.object({
    title: z.string().min(1).max(200).nullish().default(null),
    completed: z.boolean().nullish().default(null)
});

/**
 * TodoDeleteResponse
 */
export const zTodoDeleteResponse = z.object({
    id: z.uuid(),
    txid: z.int()
});

/**
 * Successful response
 */
export const zGetReadyzResponse = zHealthResponse;

/**
 * Successful response
 */
export const zGetHealthzResponse = zHealthResponse;

export const zPostApiV1TodosBody = zTodoCreate;

/**
 * Successful response
 */
export const zPostApiV1TodosResponse = zTodoMutationResponse;

/**
 * The fixed Todo shape stream from Electric.
 */
export const zGetApiV1SyncTodosResponse = z.array(z.unknown());

export const zDeleteApiV1TodosByTodoIdHeaders = z.object({
    'If-Match': z.string()
});

export const zDeleteApiV1TodosByTodoIdPath = z.object({
    todo_id: z.string()
});

/**
 * Successful response
 */
export const zDeleteApiV1TodosByTodoIdResponse = zTodoDeleteResponse;

export const zPatchApiV1TodosByTodoIdBody = zTodoPatch;

export const zPatchApiV1TodosByTodoIdHeaders = z.object({
    'If-Match': z.string()
});

export const zPatchApiV1TodosByTodoIdPath = z.object({
    todo_id: z.string()
});

/**
 * Successful response
 */
export const zPatchApiV1TodosByTodoIdResponse = zTodoMutationResponse;